add entry

+++ entity/logEntry.go
@@ -20,5 +20,5 @@ type LogEntryEntity struct {
 	UserId      primitive.ObjectID `json:"userId" bson:"userId"`
 	CreatedAt   time.Time          `json:"createdAt" bson:"createdAt"`
 	UpdatedAt   time.Time          `json:"updatedAt" bson:"updatedAt"`
-	NumberCount int32              `json:"numberCount" bson:"numberCount"`
+	NumberCount int64              `json:"numberCount" bson:"numberCount"`
 }
